Move extractTitleFromIdentifier to the url module
Deriving a title from a doc:// identifier is path parsing, and it was the only reason the HIG renderer imported from the reference renderer. The two domains are now independent.
