+++ src/lib/url.ts
@@ -53,3 +53,38 @@ export function isValidAppleDocUrl(url: string): boolean {
 
   return url.startsWith(APPLE_DOC_BASE_URL)
 }
+
+/**
+ * Derives a human-readable title from the last component of a documentation
+ * path or `doc://` identifier.
+ *
+ * @param identifier - The identifier or path (e.g., `doc://com.apple.SwiftUI/documentation/SwiftUI/View/body-8kl5o`)
+ * @returns The title, with any disambiguation suffix removed (e.g., `body`)
+ */
+export function extractTitleFromIdentifier(identifier: string): string {
+  const parts = identifier.split("/")
+  const lastPart = parts[parts.length - 1]
+
+  // Handle disambiguation suffixes (e.g., "body-8kl5o" -> "body", "init(exactly:)-63925" -> "init(exactly:)")
+  const disambiguationMatch = lastPart.match(/^(.+?)(?:-\w+)?$/)
+  if (disambiguationMatch) {
+    const baseName = disambiguationMatch[1]
+
+    // If it looks like a method signature (contains parentheses), preserve it
+    if (baseName.includes("(") && baseName.includes(")")) {
+      return baseName
+    }
+
+    // For simple identifiers, convert camelCase to readable format
+    return baseName
+      .replace(/([a-z])([A-Z])/g, "$1 $2")
+      .replace(/\s+/g, " ")
+      .trim()
+  }
+
+  // Fallback: convert camelCase to readable format
+  return lastPart
+    .replace(/([a-z])([A-Z])/g, "$1 $2")
+    .replace(/\s+/g, " ")
+    .trim()
+}
